Add support for specify canonical source repositories for P2 Artifacts
As there is no canonical source for P2 like maven-central it is hard to
get "right" location for an artifact. It could even be that an internal
proxy is sued and then the actual (public) source is hidden.

This now adds a new option -p2sources where one can specify a list of
artifact repositories that are considered as the canonical source for
the product to be build, e.g. this could be a specific eclipse release
the product was build against, but maybe many (e.g. orbit, m2e, or other
third party providers) that are searched for an artifact in given order
and if one is found this url is used.